+++ packages/google-auth/testing/constraints-3.14.txt
@@ -1,4 +1,10 @@
-# Lower-bound constraints for Python 3.14 core dependencies.
-# Used during CI unit tests to verify minimum supported package versions.
+# This constraints file is used to check that lower bounds
+# are correct in setup.py
+# List *all* library dependencies and extras in this file.
+# Pin the version to the lower bound.
+#
+# e.g., if setup.py has "foo >= 1.14.0, < 2.0.0dev",
+# Then this file should have foo==1.14.0
 pyasn1-modules==0.2.1
 cryptography==41.0.5
+aiohttp==3.9.0
